[Eir]: Phase 1 - Prequisites - updated to MTA:BLUE revision 6096 (WIP) - moved license plate logic (CMP -> CGame) - moved clothes cache logic (CMP -> CGame) - fixed memory writes (using MemPut API now) - updated FileSystem implementation to 1.2 (+ added project)
So, here is the new plan.
Finally, I want to make this repository my main workspace. To do that I need to get this repository up to date and move over my code from my Assembla SVN. This process is going to take 2 to three months; you should see me commit more often.

Quick overview:
Phase 1: Prerequisites - getting ready for the merge
Phase 2: Implementation Update - bringing in Assembla updated code to old Eir modules
Phase 3: Research and Stability testing - seeing how new copes with old
Phase 4: Module Import - gettings new features into Eir

BUILD STATUS: not-ready

+++ forks/eir/MTA10/core/CVertexStreamBoundingBoxManager.cpp
@@ -177,8 +177,8 @@ bool CVertexStreamBoundingBoxManager::GetVertexStreamBoundingBox ( SCurrentState
         pStreamBoundsInfo = CreateStreamBoundsInfo ( state );
 
     // Calc area we are going to use
-    WORD viMinBased = state.args.MinVertexIndex + state.args.BaseVertexIndex;
-    WORD viMaxBased = state.args.MinVertexIndex + state.args.NumVertices + state.args.BaseVertexIndex;
+    uint viMinBased = state.args.MinVertexIndex + state.args.BaseVertexIndex;
+    uint viMaxBased = state.args.MinVertexIndex + state.args.NumVertices + state.args.BaseVertexIndex;
 
     uint ReadOffsetStart = viMinBased * state.stream.Stride + state.stream.OffsetInBytes;
     uint ReadOffsetSize = ( viMaxBased - viMinBased ) * state.stream.Stride;
